Fix duplicate image bug on resize, work in progress
The bug basically happens when observer.next() is called _after_ observer.error() is called, and my fix is to make only one or the other called.  However, the code for this section hasn't been updated in 7 months, so I'm going to investigate further on 1.2 (working branch) to see if it behaves the same.