Fix bug in log service with http requests not firing immediately
Solution is to subscribe to the Observable immediately (which is required to trigger the observable's actions).  So, I unwrap the observable and provide two callbacks as a simple solution, while also returning the subscription object directly too.  Currently the subscription isn't used anywhere in the app, nor was it prior to these changes, so my work here affects no implementations while enabling broad support for consuming the logging service.