Upgraded to Angular ^5.0.0 with happy path working.
Most of the difficult work had to do with the upgrade from Http to HttpClient. In particular, HttpClient now tries to automatically parse the response as JSON, but since most of our responses were plaintext this meant errors on every request.  As such, I've had to explicitly declare most responses as text (including XML, which we take as text then parse afterwards).