Fix issue with childrenCount validation, input type
childrenCount was a text field expecting a numeric input. Downstream code was
casting the string value as a number in multiple locations (why not do it at the boundary?)
and the field's type was set to number already. This seemed like a potential vector for bugs.

The validator was also incomplete; it was still possible to enter invalid values. This has
been fixed by using a regular expression which only allows numbers within the allowed range.

Some other minor issues were resolved, such as unclear indexes used to indicate claimants
and claim types for applicants/spouses and attendant care/disability respectively. This is
also intended to alleviate issues with code duplication errors from Code Climate.