DEAM-330: Ensure radio buttons need only one click
What I did:
1) Moved the existing fix onto the other remaining common-date slots
2) Fixed a spelling error

For now I just reused the existing fix but if at some point in the
future someone decides to change the height of the error container, it
will go back to the old behavior. If it's possible to alter the
moh-common-styles repo, I think it should be changed there eventually.