DEAM-322: Fix dif. labels jumping to same select
Small change, please review when convenient.

Made for and id attrs unique. This is a change to msp-core but it is a
very small change to a component that is only used in DEAM so far.

The unique attrs mean clicking on an update label (like Correct Name)
will focus the respective dropdown, rather than all labels focusing
the first dropdown.