DEAM-348: Make "From" field mandatory if visible
This is for the Spouse Info page.

What I did:
1) Changed validation from just being falsy or truthy to not undefined
and not null, for any odd cases where the value gets reset to null etc.
2) Added checks for arrays/objects before the length of content of them
was validated. Eg:
Instead of:
`if (some property of the object exists)`
I put:
`if (the object exists && some property of the object exists)`
3) Added validation for the "From (province or country)" field
4) Removed a comment in moving-information that is no longer needed

Could still use more validation