DEAM-351: Upload failure alerts
This is a bit of a verbose message but this component is used in
multiple places and it is a fairly helpful feature.

What I did:
1) Made use of errorDocument emitter in common-file-uploader
2) Replaced checkDuplicateDocs and newImage with supportDocError
(getter + setter), supportDocErrorMsg (getter + setter), and
handleSupportDocError.
3) Use enums to dictate which messages to dynamically show

Tried across IE, Chrome, etc. and in different combinations. The
remaining situations it doesn't handle well already exist in the
current version, and those is the following:
1) If you attempt to upload a filetype that isn't one of the expected
ones (.otf for example) it will show the error as expected, but
subsequent correct uploads will not run. For some reason it stops at
`opening file dialog` in the common-file-uploader component. The issue
goes away when you rerender the component (by unchecking and rechecking
the relevant update checkbox)

2) It is limited to 50 pages for PDFs (makes sense if youre only upload
IDs and short forms) but will cut off longer PDFs without notifying the
user. It also takes long enough that the user may assume it's broken.
I intend to add a loading spinner for this as well as a restyle.

To test:
1) Upload a proper doc
2) Click the add button and upload the same doc again. There should be
an appropriate error.
3) Click remove to remove the current doc, the error should go away.
4) Add another valid document, then add a different kind of valid doc
(multipage PDF etc) and make sure that passes as well (you may have to
wait a while depending on how many pages there are)