DEAM-370: Edit review page content
This was an update requested by Ernesto. Please review when convenient

What I did:
1) Put PHN before Status In Canada and moved document count to the
bottom of the card (as per mockups). Moved school
address next to rest of school info
2) Removed 12 and 6 month dates, reasons, locations. Left only the yes
or no value (Ernesto asked for this)
3) Added mailing address for removed spouses and children, put it at the
bottom of the cards (Ernesto asked for this)
4) Fixed hasDocumentAttached and documentCount to include spouse and
children files by changing the check to be for documents
themselves rather than the fields associated with documents. This
is because children and spouses don't have all the same fields as the
AH so the documents themselves are the only way to check all three,
AH, spouse, and child (Ernesto asked for document counts to show
for all types of cards)

See attached for before and afters