DEAM-75: Update spec for account module
Please review and approve when convenient

What I did:
Updated the spec for the account module. At some point `processService`
was added to the constructor but the spec was not updated accordingly.So
it was expecting an argument for processService but not receiving one.
Tests should run as normal now