DEAM-420: Disable uploads with too high a size
To test this branch, upload a bunch of large images, over ~1.3MB after
compression, in one support document slot (enough that you'd get a
failure on current test deploy.) On this branch you should see an error
message, and the slot should be cleared. Then upload docs that are
small/few enough that it allows you to continue and removes the error,
and ensure you have the correct number of docs in the review section.
Lastly, make sure the submission is successful.